refactoring dspy configuration out of main to be reused by other implementations
