+++ main.py
@@ -8,15 +8,15 @@ from collections.abc import Sequence
 from dataclasses import dataclass
 from typing import Any
 
-import dspy
 from dotenv import load_dotenv
 from rich.console import Console
 from rich.prompt import Confirm, Prompt
 
+from dspy_runtime import configure_dspy, dspy_config_from_namespace
 from dspy_optimization import try_load_optimized_module
 from exceptions import RECOVERABLE_RUNTIME_EXCEPTIONS
 from image_gen import ImageGenerator
-from logging_config import TokenUsageCallback, setup_logging
+from logging_config import setup_logging
 from postprocessing import find_similar_sentences, format_report
 from story_modules import (
     ChapterInpaintingGenerator,
@@ -30,11 +30,6 @@ from story_modules import (
 from world_bible import WorldBible
 from world_bible_modules import WorldBibleGenerator, WorldBibleQuestionGenerator
 
-try:
-    from openinference.instrumentation.dspy import DSPyInstrumentor
-except ImportError:
-    DSPyInstrumentor = None
-
 load_dotenv()
 
 logger = logging.getLogger(__name__)
@@ -42,89 +37,11 @@ logger = logging.getLogger(__name__)
 console = Console()
 
 
-@dataclass
-class DSPyConfig:
-    """Runtime configuration for DSPy language model setup."""
-
-    model_name: str
-    api_base: str | None = None
-    api_key: str | None = None
-    max_tokens: int = 8000
-    cache: bool = True
-    memory_cache: bool = True
-    cache_dir: str | None = None
-
-
-def dspy_config_from_namespace(args: Namespace) -> DSPyConfig:
-    """Build DSPy runtime config from parsed CLI arguments."""
-    return DSPyConfig(
-        model_name=args.model,
-        api_base=args.llm_url,
-        api_key=args.api_key,
-        max_tokens=args.max_tokens,
-        cache=args.cache,
-        memory_cache=args.memory_cache,
-        cache_dir=args.cache_dir,
-    )
-
-
 def _env_flag_true(name: str, default: bool = False) -> bool:
     raw_value = os.environ.get(name)
     if raw_value is None:
         return default
     return raw_value.strip().lower() in {"1", "true", "yes", "on"}
-
-
-def configure_dspy(config: DSPyConfig) -> None:
-    """Configure DSPy LM and optional instrumentation."""
-    model_name = config.model_name
-    kwargs = {}
-    if config.api_base:
-        kwargs["api_base"] = config.api_base
-    if config.api_key is not None:
-        kwargs["api_key"] = config.api_key
-    elif "openai" in model_name.lower():
-        env_key = os.environ.get("OPENAI_API_KEY")
-        if not env_key:
-            logger.warning(
-                "OPENAI_API_KEY not found in environment variables. "
-                "Assuming mock or alternative setup."
-            )
-        else:
-            kwargs["api_key"] = env_key
-    elif "ollama" in model_name.lower():
-        pass
-        # kwargs["api_key"] = "" # Ollama typically doesn't need an API key
-
-    dspy.configure_cache(
-        enable_disk_cache=config.cache,
-        enable_memory_cache=config.memory_cache,
-        disk_cache_dir=config.cache_dir,
-    )
-
-    logger.info(
-        "Configuring DSPy model=%r max_tokens=%s cache=%s memory_cache=%s cache_dir=%r",
-        model_name,
-        config.max_tokens,
-        config.cache,
-        config.memory_cache,
-        config.cache_dir,
-    )
-    lm = dspy.LM(
-        model_name,
-        max_tokens=config.max_tokens,
-        cache=config.cache,
-        **kwargs,
-    )
-
-    if os.environ.get("LANGFUSE_PUBLIC_KEY") and DSPyInstrumentor is not None:
-        try:
-            DSPyInstrumentor().instrument()
-        except RECOVERABLE_RUNTIME_EXCEPTIONS as exc:
-            logger.warning("Langfuse DSPy callback handler unavailable: %s", exc)
-
-    callbacks = [TokenUsageCallback()] if TokenUsageCallback is not None else []
-    dspy.configure(lm=lm, callbacks=callbacks)
 
 
 def initialize_text_generators(
